Rename hooks to be more consistent
