+++ static/js/src/advantage/subscribe/react/APICalls/registerPaymentMethod.jsx
